columns down compressible option
option to toggle if the columns down are compressible or not. If the columns are compressible performs two passes on the moment distribution after pass one beam reactions on the columns are determined and the columns are shortened by PL/AE then the resulting fixed end moments created by the differential shortening of the columns are determined and another moment distribution pass is performed.