Non-Sway Reaction
As currently set up the Frame programs are assumed to be be non-sway. However the necessary restraining reaction was not listed when reactions were shown in the chart, this is now corrected and the non-sway reaction is shown. 

Simple check of sum of Fx in the older version would have shown a complimentary reaction at the beam line would be required to remain in equilibrium.